Refactor vertex and chunk data handling
Updated Buffers.cpp and Buffers.h to use GL_BYTE for vertex attributes and modified BufferObjects constructor to accept a vector of Vertex. Simplified chunk generation in Chunk.cpp and Chunk.h by removing insertBlocks and generateChunk methods, adding getVisibleBlocks, and updating the Chunk constructor. Introduced new data structures for vertex, face, and chunk data. Changed MAX_VISIBLE_BLOCKS type to size_t in Constants.h. Included <vector> header in Shape.h. Updated main.cpp to use the new Chunk constructor. Added constexpr constructors to uint8_VEC in vectors.h.

+++ Project1/Constants.h
@@ -5,7 +5,7 @@ namespace Constants {
 	constexpr int WINDOW_HEIGHT = 800;
 	constexpr int WINDOW_WIDTH = 600;
 
-	constexpr int MAX_VISIBLE_BLOCKS = 4096 / 2;
+	constexpr size_t MAX_VISIBLE_BLOCKS = 4096 / 2;
 }
 
 #endif // CONSTANTS_H
